Rename caseFactory to testDataFactory
