+++ apps/judicial-system/web-e2e/src/integration/Judge/courtRecord.spec.ts
@@ -1,5 +1,5 @@
 import { Case } from '@island.is/judicial-system/types'
-import { makeCase, makeProsecutor } from '../../fixtures/caseFactory'
+import { makeCase, makeProsecutor } from '../../fixtures/testDataFactory'
 import { intercept } from '../../utils'
 
 describe('/domur/thingbok/:id', () => {
