+++ apps/judicial-system/web-e2e/src/integration/Judge/rulingStepOne.spec.ts
@@ -1,7 +1,7 @@
 /// <reference path="../../support/index.d.ts" />
 
 import { Case } from '@island.is/judicial-system/types'
-import { makeCase } from '../../fixtures/caseFactory'
+import { makeCase } from '../../fixtures/testDataFactory'
 import { intercept } from '../../utils'
 
 describe('/domur/urskurdur/:id', () => {
