+++ apps/judicial-system/web-e2e/src/integration/Judge/rulingStepTwo.spec.ts
@@ -4,7 +4,7 @@ import {
   CaseDecision,
   CaseType,
 } from '@island.is/judicial-system/types'
-import { makeCase } from '../../fixtures/caseFactory'
+import { makeCase } from '../../fixtures/testDataFactory'
 import { intercept } from '../../utils'
 
 describe('/domur/urskurdarord/:id', () => {
